Add onCreate and postCreate to devcontainer

+++ .devcontainer/scripts/post_create.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-sudo rm -rf /var/www/html && sudo ln -s $CODESPACE_VSCODE_FOLDER/drupal /var/www/html
-
-mkdir -p /vscode/.ssh
-echo -e "$DRUPAL_SSH_PRIVATE_KEY\n" > /home/vscode/.ssh/id_rsa
-chmod 400 /home/vscode/.ssh/id_rsa
+if [ ! -f /home/vscode.ssh/id_rsa ];then
+  mkdir -p /home/vscode/.ssh
+  echo -e "$DRUPAL_SSH_PRIVATE_KEY\n" > /home/vscode/.ssh/id_rsa
+  chmod 400 /home/vscode/.ssh/id_rsa
+fi
